# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Usersdata",
    "description": "Information about users",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/users/": {
      "get": {
        "description": "Get all users",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "Create new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "first_name": {
                  "example": "any"
                },
                "last_name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Get user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "Remove existing user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 8
        },
        "email": {
          "type": "string",
          "example": "n.surname@gmail.com"
        },
        "first_name": {
          "type": "string",
          "example": "Name"
        },
        "last_name": {
          "type": "string",
          "example": "Surname"
        }
      }
    },
    "Users": {
      "type": "array",
      "items": {
        "type": "array",
        "items": {}
      }
    }
  }
}